remove MinGW dependecy on GpG4Win (gpg-agent)
The gpg-agent that ships with GpG4Win is a version 2 gpg agent and only works with gpg2.  The gpg in MinGW is gpg 1.x and it doesn't pick up this gpg-agent.

However, everything still works - you just have to enter your password for each file you decrypt.  Slightly onerous but not a huge deal breaker.

I notice there is  away to specify your own gpg versoin in the _blackbox_common.sh file but I don't really know how that works.  If someone were to specify to use gpg2 then they could probably use the agent in Gpg4Win but I don't know and don't know how to test that.